Split classes token and itoken
